requirements.txt
faiss-cpu==1.8.0
numpy==1.26.4
pandas==2.2.1
sentence-transformers==2.5.1
sentencepiece==0.2.0
